HOTT-2381: Adds ten_digit_goods_nomenclature and share behaviour
